added model valid in RMController

+++ ApiController/RMessageController.cs
@@ -23,10 +23,17 @@ namespace XforumTest.ApiController
         [HttpPost]
         public IActionResult CreateMessage(RMessageDTO dto)
         {
-            var service = new RMessageService();
-            service.Create(dto);
+            if (!ModelState.IsValid)
+            {
+                var service = new RMessageService();
+                service.Create(dto);
 
-            return Ok("create msg");
+                return Ok("create msg");
+            }
+            else
+            {
+                return BadRequest("model wrong");
+            }
         }
 
 
@@ -39,9 +46,15 @@ namespace XforumTest.ApiController
         [HttpGet("{postId}")]
         public async Task<ActionResult<List<RMessageDTO>>> GetAllMessages(Guid postId)
         {
-
-            var service = new RMessageService();
-            return await service.GetAllbyPostId(postId);
+            if (postId!=null)
+            {
+                var service = new RMessageService();
+                return await service.GetAllbyPostId(postId);
+            }
+            else
+            {
+                return BadRequest("postId is null");
+            }
         }
 
 
@@ -53,9 +66,17 @@ namespace XforumTest.ApiController
         [HttpDelete]
         public IActionResult DeleteMessages(Guid mId)
         {
-            var service = new RMessageService();
-            service.Delete(mId);
-            return Ok("delete msg");
+            if (mId != null)
+            {
+                var service = new RMessageService();
+                service.Delete(mId);
+                return Ok("delete msg");
+            }
+            else
+            {
+                return BadRequest("Id is null");
+            }
+           
         }
 
 
@@ -67,10 +88,17 @@ namespace XforumTest.ApiController
         [HttpPut]
         public IActionResult PostLikeAndDisLike(MessageLikeDto Dto)
         {
-            var service = new LikeService2();
-            service.PostLikeAndDisLike(Dto);
-
-            return Ok("delete msg");
+            if (!ModelState.IsValid)
+            {
+                var service = new LikeService2();
+                service.PostLikeAndDisLike(Dto);
+                return Ok("create msg");
+            }
+            else
+            {
+                return BadRequest("Dto wrong");
+            }
+           
         }
     
     }
